samples/password-generator/level-1-basic/nanoagent/pyproject.toml
[project]
name = "nanoagent-basic"
version = "0.1.0"
description = "Level 1: Basic single-agent example - one-shot prompt/response"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agent = "agent:main"
